refactor(bd-6pi): suffix worker entry chain modules with their condition
The RSC worker runs under react-server; the HTML worker runs under
react-client. Their shared module set (state, handleRscRender /
handleHtmlRender, hydrateRscRenderMessage, messageHandler) was
condition-neutral by filename but condition-bound at module init —
each statically imports vendor.{server,client}.js or
helpers/index.server.js. Under bd-6pi's never-mix rule that is a
neutral-importing-suffixed violation even though the runtime
condition happens to match.

Rename:
  worker/rsc/state.ts                 -> state.server.ts
  worker/rsc/handleRscRender.ts       -> handleRscRender.server.ts
  worker/rsc/hydrateRscRenderMessage  -> hydrateRscRenderMessage.server.ts
  worker/rsc/messageHandler.tsx       -> messageHandler.server.tsx
  worker/html/state.ts                -> state.client.ts
  worker/html/handleHtmlRender.ts     -> handleHtmlRender.client.ts
  worker/html/messageHandler.tsx      -> messageHandler.client.tsx

Update the worker-entry trampolines (rsc-worker.{development,production}.ts,
html-worker.{development,production}.tsx) to import the suffixed
modules. The trampolines themselves stay neutral-by-filename: they
are entry points spawned with explicit --conditions execArgv, so
their condition is set by Node's worker spawn path, not by ESM
linking — and the bd-6pi violation list does not include them.
